Adicionando exercícios Matrizes

+++ Exercicio2.java
@@ -4,13 +4,13 @@ public class Exercicio2 {
 
     public static void main(String[] args){
         Scanner scanner = new Scanner(System.in);
-        double[] valores = new double[10];
-        double pares = 0;
-        double impares = 0;
+        int[] valores = new int[10];
+        int pares = 0;
+        int impares = 0;
 
         for(int i=0; i < 10; i++){
             System.out.println("Digite um valor: ");
-            valores[i] = scanner.nextDouble();
+            valores[i] = scanner.nextInt();
 
             if(valores[i] % 2 == 0){
                 pares++;
