+++ Exercicio5.java
@@ -5,13 +5,13 @@ public class Exercicio5{
     public static void main(String[] args){
 
         Scanner scanner = new Scanner(System.in);
-        double[] notas = new double[15];
-        double soma = 0;
-        double mediaGeral;
+        int[] notas = new int[15];
+        int soma = 0;
+        int mediaGeral;
 
         for(int i = 0; i < 15; i++){
             System.out.println("Digite a nota do aluno " + (i + 1));
-            notas[i] = scanner.nextDouble();
+            notas[i] = scanner.nextInt();
             soma += notas[i];
         }
         mediaGeral = soma / notas.length;
